Add support and tests for process models

+++ edu.ksu.cis.projects.mdcf.aadl-translator-test/src/test/java/edu/ksu/cis/projects/mdcf/aadltranslator/test/arch/PortModelTests.java
@@ -51,7 +51,7 @@ public class PortModelTests {
 
 	@Test
 	public void testPortType() {
-		assertEquals("Integer", devSendPort.getType());
+		assertEquals("Double", devSendPort.getType());
 		assertEquals("Object", procSendPort.getType());
 	}
